mejora de formularios

+++ routes/web.php
@@ -59,50 +59,50 @@ Route::get('/auditor/configuracion', [ConfiguracionController::class, 'index'])-
 
 //Rutas Administrador//
 //Listar todos los usuarios
-Route::get('/admin/gestionar-usuarios', [UserController::class, 'index'])->name('usuarios.index')->middleware('rol:administrador');
+Route::get('/admin/gestionar-usuarios', [UserController::class, 'index'])->name('admin.usuarios.index')->middleware('rol:administrador');
 //formulario para crear un nuevo usuario
-Route::get('/admin/usuarios/crear-usuario', [UserController::class, 'create'])->name('usuarios.create')->middleware('rol:administrador');
+Route::get('/admin/usuarios/crear-usuario', [UserController::class, 'create'])->name('admin.usuarios.create')->middleware('rol:administrador');
 // Guardar nuevo usuario
-Route::post('/admin/guardar-usuario', [UserController::class, 'store'])->name('usuarios.store')->middleware('rol:administrador');
+Route::post('/admin/guardar-usuario', [UserController::class, 'store'])->name('admin.usuarios.store')->middleware('rol:administrador');
 // Mostrar detalles de un usuaurio
-Route::get('/admin/detallar-usuario/{user}', [UserController::class, 'show'])->name('usuarios.show')->middleware('rol:administrador');
+Route::get('/admin/detallar-usuario/{user}', [UserController::class, 'show'])->name('admin.usuarios.show')->middleware('rol:administrador');
 //formulario para editar un usuario
-Route::get('/admin/usuarios/{user}/editar-usuario', [UserController::class, 'edit'])->name('usuarios.edit')->middleware('rol:administrador');
+Route::get('/admin/usuarios/{user}/editar-usuario', [UserController::class, 'edit'])->name('admin.usuarios.edit')->middleware('rol:administrador');
 // Actualizar un usuario
-Route::put('/admin/actualizar-usuario/{user}', [UserController::class, 'update'])->name('usuarios.update')->middleware('rol:administrador');
+Route::put('/admin/actualizar-usuario/{user}', [UserController::class, 'update'])->name('admin.usuarios.update')->middleware('rol:administrador');
 // Eliminar un usuario
-Route::delete('/admin/eliminar-usuario/{user}', [UserController::class, 'destroy'])->name('usuarios.destroy')->middleware('rol:administrador');
+Route::delete('/admin/eliminar-usuario/{user}', [UserController::class, 'destroy'])->name('admin.usuarios.destroy')->middleware('rol:administrador');
 
 // Listar todas las sedes
-Route::get('/admin/gestionar-sedes', [SedeController::class, 'index'])->name('sedes.index')->middleware('rol:administrador');
+Route::get('/admin/gestionar-sedes', [SedeController::class, 'index'])->name('admin.sedes.index')->middleware('rol:administrador');
 //formulario para crear una nueva sede
-Route::get('/admin/sedes/crear-sede', [SedeController::class, 'create'])->name('sedes.create')->middleware('rol:administrador');
+Route::get('/admin/sedes/crear-sede', [SedeController::class, 'create'])->name('admin.sedes.create')->middleware('rol:administrador');
 // Guardar nueva sede
-Route::post('/admin/guardar-sede', [SedeController::class, 'store'])->name('sedes.store')->middleware('rol:administrador');
+Route::post('/admin/guardar-sede', [SedeController::class, 'store'])->name('admin.sedes.store')->middleware('rol:administrador');
 // Mostrar detalles de una sede
-Route::get('/admin/detallar-sede/{sede}', [SedeController::class, 'show'])->name('sedes.show')->middleware('rol:administrador');
+Route::get('/admin/detallar-sede/{sede}', [SedeController::class, 'show'])->name('admin.sedes.show')->middleware('rol:administrador');
 //formulario para editar una sede
-Route::get('/admin/sedes/{sede}/editar-sede', [SedeController::class, 'edit'])->name('sedes.edit')->middleware('rol:administrador');
+Route::get('/admin/sedes/{sede}/editar-sede', [SedeController::class, 'edit'])->name('admin.sedes.edit')->middleware('rol:administrador');
 // Actualizar una sede
-Route::put('/admin/actualizar-sede/{sede}', [SedeController::class, 'update'])->name('sedes.update')->middleware('rol:administrador');
+Route::put('/admin/actualizar-sede/{sede}', [SedeController::class, 'update'])->name('admin.sedes.update')->middleware('rol:administrador');
 // Eliminar una sede
-Route::delete('/admin/eliminar-sede/{sede}', [SedeController::class, 'destroy'])->name('sedes.destroy')->middleware('rol:administrador');
+Route::delete('/admin/eliminar-sede/{sede}', [SedeController::class, 'destroy'])->name('admin.sedes.destroy')->middleware('rol:administrador');
 
 
 // Listar todas las tareas
-Route::get('/admin/gestionar-tareas', [TareaController::class, 'index'])->name('tareas.index')->middleware('rol:administrador');
+Route::get('/admin/gestionar-tareas', [TareaController::class, 'index'])->name('admin.tareas.index')->middleware('rol:administrador');
 //formulario para crear una nueva tarea
-Route::get('/admin/tareas/crear-tarea', [TareaController::class, 'create'])->name('tareas.create')->middleware('rol:administrador');
-// Guardar nueva sede
-Route::post('/admin/guardar-tarea', [TareaController::class, 'store'])->name('tareas.store')->middleware('rol:administrador');
+Route::get('/admin/tareas/crear-tarea', [TareaController::class, 'create'])->name('admin.tareas.create')->middleware('rol:administrador');
+// Guardar nueva tarea
+Route::post('/admin/guardar-tarea', [TareaController::class, 'store'])->name('admin.tareas.store')->middleware('rol:administrador');
 // Mostrar detalles de una tarea
-Route::get('/admin/detallar-tarea/{tarea}', [TareaController::class, 'show'])->name('tareas.show')->middleware('rol:administrador',);
+Route::get('/admin/detallar-tarea/{tarea}', [TareaController::class, 'show'])->name('admin.tareas.show')->middleware('rol:administrador',);
 //formulario para editar una tarea
-Route::get('/admin/tareas/{tarea}/editar-tarea', [TareaController::class, 'edit'])->name('tareas.edit')->middleware('rol:administrador');
+Route::get('/admin/tareas/{tarea}/editar-tarea', [TareaController::class, 'edit'])->name('admin.tareas.edit')->middleware('rol:administrador');
 // Actualizar un tarea
-Route::put('/admin/actualizar-tarea/{tarea}', [TareaController::class, 'update'])->name('tareas.update')->middleware('rol:administrador');
+Route::put('/admin/actualizar-tarea/{tarea}', [TareaController::class, 'update'])->name('admin.tareas.update')->middleware('rol:administrador');
 // Eliminar una tarea
-Route::delete('/admin/eliminar-tarea/{tarea}', [TareaController::class, 'destroy'])->name('tareas.destroy')->middleware('rol:administrador');
+Route::delete('/admin/eliminar-tarea/{tarea}', [TareaController::class, 'destroy'])->name('admin.tareas.destroy')->middleware('rol:administrador');
 
 
 //Rutas Tecnico//
@@ -110,4 +110,35 @@ Route::delete('/admin/eliminar-tarea/{tarea}', [TareaController::class, 'destroy
 // Visualizar tarea asignadas
 Route::get('/tecnico/gestionar-tareas', [TareaController::class, 'indexTecnico'])->name('tecnico.tareas.index')->middleware('rol:tecnico');
 Route::get('/tecnico/visualizar-tarea/{tarea}', [TareaController::class,'verTarea'])->name('tecnico.tareas.show') ->middleware('rol:tecnico');
+// Mostrar el formulario
+Route::get('/tecnico/resolucion-tarea/{tarea}', [TareaController::class, 'formularioResolucion'])->name('tecnico.tareas.resolucion');
+// Procesar el formulario (subir resolución)
+Route::post('/tecnico/guardar-resolucion/{tarea}', [TareaController::class, 'subirResolucion'])->name('tecnico.tareas.guardarResolucion');
 
+
+
+//Rutas Encargado
+
+//Visualizar tareas creadas
+Route::get('/encargado/tareas/gestionar-tareas', [TareaController::class, 'indexEncargado'])->name('encargado.tareas.listadoTareas')->middleware('rol:encargado');
+//Crear Tarea
+Route::get('/encargado/tareas/crear-tarea', [TareaController::class, 'crearTareaEncargado'])->name('encargado.tareas.crearTarea')->middleware('rol:encargado');
+// Guardar nueva tarea
+Route::post('/encargado/tareas/guardar-tarea', [TareaController::class, 'storeEncargado'])->name('encargado.tareas.guardarTarea')->middleware('rol:encargado');
+//formulario para editar una tarea
+Route::get('/encargado/tareas/{tarea}/editar-tarea', [TareaController::class,'editEncargado'])->name('encargado.tareas.editarTarea') ->middleware('rol:encargado');
+// Mostrar detalles de una tarea
+Route::get('/encargado/ver-tarea/{tarea}', [TareaController::class,'showEncargado'])->name('encargado.tareas.verTarea') ->middleware('rol:encargado');
+// Actualizar un tarea
+Route::put('/encargado/actualizar-tarea/{tarea}', [TareaController::class, 'updateEncargado'])->name('encargado.tareas.updateTarea')->middleware('rol:encargado');
+// Eliminar una tarea
+Route::delete('/encargado/borrar-tarea/{tarea}', [TareaController::class, 'destroyEncargado'])->name('encargado.tareas.destroy')->middleware('rol:encargado');
+
+
+//Rutas Sedes
+//Listado de Sedes
+Route::get('/auditor/sedes/listado-sedes', [SedeController::class, 'indexAuditor'])->name('auditor.sedes.listadoSedes')->middleware('rol:auditor');
+//Visualizar Sede
+Route::get('/auditor/detallar-sede/{sede}', [SedeController::class, 'showAuditor'])->name('auditor.sedes.verSede')->middleware('rol:auditor');
+//Visualizar Tareas Finalizadas
+Route::get('/auditor/tareas/visualizar-tareas', [TareaController::class, 'indexFinalizadasAuditor'])->name('auditor.tareas.finalizadas')->middleware('rol:auditor');
